actually move entry point to a module instead
