+++ .github/workflows/run_tests.yml
@@ -16,4 +16,4 @@ jobs:
       run: docker network create tasks
       
     - name: Run tests
-      run: docker run -v /var/run/docker.sock:/var/run/docker.sock cowait/task python -m pytest
+      run: docker run -v /var/run/docker.sock:/var/run/docker.sock --workdir /var/cowait cowait/task python -m pytest
