+++ Dockerfile
@@ -8,8 +8,8 @@ RUN apt-get update && \
     build-essential ca-certificates
 
 # create working directory
-RUN mkdir -p /app/context
-WORKDIR /app
+RUN mkdir /var/cowait && mkdir /var/task
+WORKDIR /var/cowait
 
 # install python requirements
 COPY requirements.txt .
@@ -17,11 +17,11 @@ RUN pip install -r requirements.txt
 
 # copy minimum set of files to install pip package
 COPY setup.py README.md ./
-COPY bin /app/bin
+COPY bin ./bin
 RUN pip install -e .
 
 # copy code
 COPY . .
 
-WORKDIR /app/context
-CMD [ "cowait", "exec" ]
+WORKDIR /var/task
+CMD [ "python3", "-m", "cowait.exec" ]
